+++ site/dat/docs/Changelog.md
@@ -1,5 +1,7 @@
 # Changelog
-
+## Next version
+  - implemented graceful shutdown for JMeter
+  
 ## 0.3.9 (next)
  - allow dumping final stats in Jenkins-consumable format